Rename IdGenerators to StronglyTypedIds in SharedKernel.Tests to match SharedKernel
